backend/tpdb-backend-mono/src/authentication/user/entities/user.entity.ts
import { Column, Entity, PrimaryGeneratedColumn, Unique } from 'typeorm';
import * as bcrypt from 'bcrypt';

@Entity()
@Unique(['email'])
export class User {
  @PrimaryGeneratedColumn('uuid')
  id: string;

  @Column()
  username: string;

  @Column()
  email: string;

  @Column()
  password: string;

  @Column()
  salt: string;

  async validatePassword(password: string): Promise<boolean> {
    if (password) {
      const hash = await bcrypt.hash(password, this.salt);
      return hash === this.password;
    }
    return Promise.resolve(false);
  }
}
